extract color utils into fou
