+++ app/packages/looker/src/constants.ts
@@ -29,14 +29,6 @@ export const DATE_TIME = "DateTime";
 export const SELECTION_TEXT =
   "Click to select sample, Shift+Click to select a range";
 
-export const BIG_ENDIAN = (() => {
-  let buf = new ArrayBuffer(4);
-  let u32data = new Uint32Array(buf);
-  let u8data = new Uint8Array(buf);
-  u32data[0] = 0xcafebabe;
-  return u8data[0] === 0xca;
-})();
-
 export const MOMENT_CLASSIFICATIONS = [
   CLASSIFICATION,
   CLASSIFICATIONS,
